requirements.txt
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.1
flask==3.0.3
importlib-metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
plotly==5.22.0
pluggy==1.5.0
py3-wkhtmltopdf==0.4.1
pytest==8.2.2
tenacity==8.3.0
tomli==2.0.1
werkzeug==3.0.3
zipp==3.19.2
